--- components/App.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import Panel from './Panel'
-import CyReact from '../containers/CyReact'
-import EntityPanel from '../containers/EntityPanel'
-
-const App = () => (
-	<div>
-		<Panel/>
-		<CyReact containerId='react'/>
-    <EntityPanel />
-	</div>
-)
-
-export default App
