+++ cy_canvas/reducers/cyStyle.js
@@ -1,4 +1,4 @@
-import DefaultVisualStyle from '../css/DefaultVisualStyle'
+import DefaultVisualStyle from '../../css/DefaultVisualStyle'
 
 const cyStyle = (state = DefaultVisualStyle, action) => {
   switch(action.type) {
