+++ entity_panel/components/Entity.js
