+++ entity_panel/components/EntityName.js
