+++ entity_panel/components/SynonymList.js
