+++ index.js
@@ -2,8 +2,8 @@ import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
 import { createStore } from 'redux'
-import voiceUIApp from './reducers'
-import App from './components/App'
+import voiceUIApp from './common/reducers'
+import App from './common/App'
 
 let store = createStore(voiceUIApp)
