+++ property_panel/reducers/intentPropertiesPanel.js
